% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{block_matrix_mult}
\alias{block_matrix_mult}
\title{Efficient block matrix multiplication}
\usage{
block_matrix_mult(X, Y, block_size = 500)
}
\arguments{
\item{X}{First matrix}

\item{Y}{Second matrix}

\item{block_size}{Block size for computation (default: 500)}
}
\value{
Matrix product X %*% Y
}
\description{
Efficient block matrix multiplication
}
