% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{cache_stats}
\alias{cache_stats}
\title{Get cache statistics}
\usage{
cache_stats(cache_env = NULL)
}
\arguments{
\item{cache_env}{Cache environment}
}
\value{
List with cache statistics
}
\description{
Get cache statistics
}
