% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiobjective.R
\name{calculate_crowding_distance}
\alias{calculate_crowding_distance}
\title{Calculate crowding distance for diversity preservation}
\usage{
calculate_crowding_distance(fitness)
}
\arguments{
\item{fitness}{Fitness matrix for a single front}
}
\value{
Vector of crowding distances
}
\description{
Calculate crowding distance for diversity preservation
}
