% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga-operators.R
\name{calculate_diversity}
\alias{calculate_diversity}
\title{Calculate population diversity}
\usage{
calculate_diversity(population)
}
\arguments{
\item{population}{List of solutions in the population}
}
\value{
Diversity measure (0 to 1, higher = more diverse)
}
\description{
Calculate population diversity
}
