% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiobjective.R
\name{check_dominance}
\alias{check_dominance}
\title{Check dominance between two solutions}
\usage{
check_dominance(solution1, solution2, criteria_types)
}
\arguments{
\item{solution1}{First solution fitness}

\item{solution2}{Second solution fitness}

\item{criteria_types}{Optimization types}
}
\value{
1 if solution1 dominates, -1 if solution2 dominates, 0 if non-dominated
}
\description{
Check dominance between two solutions
}
