% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_memory_usage}
\alias{check_memory_usage}
\title{Check memory usage and warn for large computations}
\usage{
check_memory_usage(n_individuals, n_variables, operation = "matrix_mult")
}
\arguments{
\item{n_individuals}{Number of individuals}

\item{n_variables}{Number of variables/markers}

\item{operation}{Type of operation ("matrix_mult", "eigen", "svd", "chol")}
}
\value{
NULL, but issues warnings for large computations
}
\description{
Check memory usage and warn for large computations
}
