% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear computation cache}
\usage{
clear_cache(cache_env = NULL)
}
\arguments{
\item{cache_env}{Cache environment to clear}
}
\description{
Clear computation cache
}
