% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-operations.R
\name{compute_amatrix}
\alias{compute_amatrix}
\title{STPGA Matrix Operations}
\usage{
compute_amatrix(M, pieces = 10, mc.cores = 1)
}
\arguments{
\item{M}{Marker matrix coded as -1,0,1 (individuals x markers)}

\item{pieces}{Number of pieces to split computation (default: 10)}

\item{mc.cores}{Number of cores for parallel processing (default: 1)}
}
\value{
Combined genomic relationship matrix (A matrix)
}
\description{
Matrix utilities for genetic relationship computations
}
\author{
Deniz Akdemir
Compute A matrix in pieces for large marker datasets
}
