% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiobjective.R
\name{compute_convergence_metrics}
\alias{compute_convergence_metrics}
\title{Compute convergence metrics for multi-objective optimization}
\usage{
compute_convergence_metrics(fitness)
}
\arguments{
\item{fitness}{Archive fitness matrix}
}
\value{
Vector of convergence metrics
}
\description{
Compute convergence metrics for multi-objective optimization
}
