% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{compute_population_stats}
\alias{compute_population_stats}
\title{Population statistics computation}
\usage{
compute_population_stats(population, fitness)
}
\arguments{
\item{population}{List of solutions}

\item{fitness}{Vector of fitness values}
}
\value{
List with population statistics
}
\description{
Population statistics computation
}
