% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming-standards.R
\name{create_function_alias}
\alias{create_function_alias}
\title{Create function alias for backward compatibility}
\usage{
create_function_alias(old_name, new_name, envir = .GlobalEnv)
}
\arguments{
\item{old_name}{Old function name}

\item{new_name}{New standardized function name}

\item{envir}{Environment to create alias in}
}
\description{
Create function alias for backward compatibility
}
