% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{create_progress_bar}
\alias{create_progress_bar}
\title{Performance Optimization Functions}
\usage{
create_progress_bar(total, title = "Processing", verbose = TRUE)
}
\arguments{
\item{total}{Total number of iterations}

\item{title}{Title for the progress bar}

\item{verbose}{Whether to show progress bar}
}
\value{
Progress bar object or NULL
}
\description{
Functions to improve computational performance for large datasets
}
\author{
Deniz Akdemir
Add progress bar for long-running computations
}
