% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga-operators.R
\name{crowding_replacement}
\alias{crowding_replacement}
\title{Crowding replacement for diversity preservation}
\usage{
crowding_replacement(
  offspring,
  offspring_fitness,
  population,
  population_fitness,
  crowding_factor = 3
)
}
\arguments{
\item{offspring}{New offspring solutions}

\item{offspring_fitness}{Fitness of offspring}

\item{population}{Current population}

\item{population_fitness}{Current population fitness}

\item{crowding_factor}{Number of candidates to compare for replacement}
}
\value{
Updated population after crowding replacement
}
\description{
Crowding replacement for diversity preservation
}
