% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{determine_optimal_batch_size}
\alias{determine_optimal_batch_size}
\title{Determine optimal batch size for evaluation}
\usage{
determine_optimal_batch_size(n_solutions, mc.cores)
}
\arguments{
\item{n_solutions}{Number of solutions}

\item{mc.cores}{Number of cores}
}
\value{
Optimal batch size
}
\description{
Determine optimal batch size for evaluation
}
