% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiobjective.R
\name{determine_primary_matrix}
\alias{determine_primary_matrix}
\title{Determine primary matrix for optimization}
\usage{
determine_primary_matrix(Pcs, Dist, Kernel)
}
\arguments{
\item{Pcs}{Principal components matrix}

\item{Dist}{Distance matrix}

\item{Kernel}{Kernel matrix}
}
\value{
Primary matrix to use for optimization
}
\description{
Determine primary matrix for optimization
}
