% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{distance_criterion}
\alias{distance_criterion}
\title{Unified distance criterion function}
\usage{
distance_criterion(
  train,
  test,
  distance_matrix,
  criterion = c("max_to_test", "mean_to_test", "neg_min_internal", "neg_mean_internal"),
  lambda = NULL,
  C = NULL
)
}
\arguments{
\item{train}{Training set individual names}

\item{test}{Test set individual names}

\item{distance_matrix}{Distance matrix}

\item{criterion}{Type of distance criterion}

\item{lambda}{Ridge parameter (for compatibility)}

\item{C}{Contrast matrix (for compatibility)}
}
\value{
Distance criterion value
}
\description{
Unified distance criterion function
}
