% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{distance_internal_min}
\alias{distance_internal_min}
\title{Negative minimum distance within training set}
\usage{
distance_internal_min(
  train,
  test = NULL,
  distance_matrix,
  lambda = NULL,
  C = NULL
)
}
\arguments{
\item{train}{Training set individual names}

\item{test}{Test set individual names (unused, kept for compatibility)}

\item{distance_matrix}{Distance matrix}

\item{lambda}{Ridge parameter (unused, kept for compatibility)}

\item{C}{Contrast matrix (unused, kept for compatibility)}
}
\value{
Negative minimum internal distance
}
\description{
Negative minimum distance within training set
}
