% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{diversity_summary}
\alias{diversity_summary}
\title{Population diversity summary}
\usage{
diversity_summary(solutions, method = "jaccard")
}
\arguments{
\item{solutions}{List of solutions}

\item{method}{Distance method for calculation}
}
\value{
List with diversity statistics
}
\description{
Population diversity summary
}
