% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming-standards.R
\name{document_parameter}
\alias{document_parameter}
\title{Generate consistent parameter documentation}
\usage{
document_parameter(param_name, param_type, default_value = NULL, description)
}
\arguments{
\item{param_name}{Parameter name}

\item{param_type}{Parameter type}

\item{default_value}{Default value}

\item{description}{Parameter description}
}
\value{
Formatted roxygen2 parameter documentation
}
\description{
Generate consistent parameter documentation
}
