% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga-operators.R
\name{elite_selection}
\alias{elite_selection}
\title{Elite selection}
\usage{
elite_selection(population, fitness, n_select)
}
\arguments{
\item{population}{List of solutions}

\item{fitness}{Vector of fitness values}

\item{n_select}{Number of elites to select}
}
\value{
Indices of selected elite individuals
}
\description{
Elite selection
}
