% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiobjective.R
\name{environmental_selection}
\alias{environmental_selection}
\title{Environmental selection for NSGA-II}
\usage{
environmental_selection(fitness, ranks, target_size)
}
\arguments{
\item{fitness}{Combined fitness matrix}

\item{ranks}{Dominance ranks}

\item{target_size}{Target population size}
}
\value{
Indices of selected solutions
}
\description{
Environmental selection for NSGA-II
}
