% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiobjective.R
\name{evaluate_multiobjective_population}
\alias{evaluate_multiobjective_population}
\title{Evaluate population for multiple objectives}
\usage{
evaluate_multiobjective_population(
  population,
  P,
  test,
  criteria,
  lambda,
  C,
  mc.cores
)
}
\arguments{
\item{population}{List of solutions}

\item{P}{Primary matrix}

\item{test}{Test set}

\item{criteria}{Vector of criteria names}

\item{lambda}{Ridge parameter}

\item{C}{Contrast matrix}

\item{mc.cores}{Number of cores}
}
\value{
Matrix of fitness values (solutions x criteria)
}
\description{
Evaluate population for multiple objectives
}
