% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluate_population_smart}
\alias{evaluate_population_smart}
\title{Smart population evaluation with adaptive batch processing}
\usage{
evaluate_population_smart(
  population,
  P,
  test = NULL,
  criterion = "PEVMEAN2",
  lambda = 1e-06,
  C = NULL,
  K = NULL,
  Vg = NULL,
  Ve = NULL,
  mc.cores = 1,
  batch_size = NULL
)
}
\arguments{
\item{population}{List of solutions}

\item{P}{Prediction matrix}

\item{test}{Test set names}

\item{criterion}{Criterion name}

\item{lambda}{Ridge parameter}

\item{C}{Contrast matrix}

\item{K}{Kinship matrix}

\item{Vg}{Genetic variance}

\item{Ve}{Error variance}

\item{mc.cores}{Number of cores}

\item{batch_size}{Batch size for processing (default: NULL for automatic)}
}
\value{
Vector of fitness values
}
\description{
Smart population evaluation with adaptive batch processing
}
