% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluate_solution_batch}
\alias{evaluate_solution_batch}
\title{Evaluate a batch of solutions efficiently}
\usage{
evaluate_solution_batch(
  solutions,
  P,
  test,
  criterion,
  lambda,
  C,
  K,
  Vg,
  Ve,
  mc.cores
)
}
\arguments{
\item{solutions}{List of solutions in the batch}

\item{P}{Prediction matrix}

\item{test}{Test set}

\item{criterion}{Criterion name}

\item{lambda}{Ridge parameter}

\item{C}{Contrast matrix}

\item{K}{Kinship matrix}

\item{Vg}{Genetic variance}

\item{Ve}{Error variance}

\item{mc.cores}{Number of cores}
}
\value{
Vector of fitness values
}
\description{
Evaluate a batch of solutions efficiently
}
