% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiobjective.R
\name{extract_pareto_front}
\alias{extract_pareto_front}
\title{Extract final Pareto front}
\usage{
extract_pareto_front(archive, fitness, criteria_types)
}
\arguments{
\item{archive}{Archive of solutions}

\item{fitness}{Archive fitness matrix}

\item{criteria_types}{Optimization types}
}
\value{
List with Pareto solutions and fitness
}
\description{
Extract final Pareto front
}
