% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiobjective.R
\name{find_non_dominated}
\alias{find_non_dominated}
\title{Find non-dominated solutions}
\usage{
find_non_dominated(fitness, criteria_types)
}
\arguments{
\item{fitness}{Fitness matrix}

\item{criteria_types}{Optimization types}
}
\value{
Indices of non-dominated solutions
}
\description{
Find non-dominated solutions
}
