% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga-operators.R
\name{fitness_sharing}
\alias{fitness_sharing}
\title{Fitness sharing for diversity preservation}
\usage{
fitness_sharing(fitness, population, sharing_radius = 0.1, alpha = 1)
}
\arguments{
\item{fitness}{Vector of fitness values}

\item{population}{List of solutions}

\item{sharing_radius}{Radius for fitness sharing (default: 0.1)}

\item{alpha}{Sharing function exponent (default: 1)}
}
\value{
Adjusted fitness values
}
\description{
Fitness sharing for diversity preservation
}
