% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming-standards.R
\name{generate_doc_template}
\alias{generate_doc_template}
\title{Generate standard function documentation template}
\usage{
generate_doc_template(func_name, purpose, category = "General")
}
\arguments{
\item{func_name}{Function name}

\item{purpose}{Function purpose}

\item{category}{Function category}
}
\value{
Roxygen2 documentation template
}
\description{
Generate standard function documentation template
}
