% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiobjective.R
\name{generate_multiobjective_offspring}
\alias{generate_multiobjective_offspring}
\title{Generate offspring for multi-objective optimization}
\usage{
generate_multiobjective_offspring(
  parents,
  candidates,
  ntoselect,
  n_offspring,
  mutprob,
  mutintensity,
  mc.cores
)
}
\arguments{
\item{parents}{Parent solutions}

\item{candidates}{Candidate pool}

\item{ntoselect}{Solution size}

\item{n_offspring}{Number of offspring to generate}

\item{mutprob}{Mutation probability}

\item{mutintensity}{Mutation intensity}

\item{mc.cores}{Number of cores}
}
\value{
List of offspring solutions
}
\description{
Generate offspring for multi-objective optimization
}
