% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-operations.R
\name{genomic_relationship_matrix}
\alias{genomic_relationship_matrix}
\title{Compute genomic relationship matrix using different methods}
\usage{
genomic_relationship_matrix(M, method = "vanraden", pieces = 10, mc.cores = 1)
}
\arguments{
\item{M}{Marker matrix coded as -1,0,1}

\item{method}{Method for computation: "vanraden", "pieces", "standard" (default: "vanraden")}

\item{pieces}{Number of pieces for piece-wise computation (default: 10)}

\item{mc.cores}{Number of cores for parallel processing (default: 1)}
}
\value{
Genomic relationship matrix
}
\description{
Compute genomic relationship matrix using different methods
}
