% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{get_optimal_cores}
\alias{get_optimal_cores}
\title{Get optimal number of cores for computation}
\usage{
get_optimal_cores(data_size, operation_type = "mixed")
}
\arguments{
\item{data_size}{Approximate size of data}

\item{operation_type}{Type of operation ("cpu", "memory", "mixed")}
}
\value{
Recommended number of cores
}
\description{
Get optimal number of cores for computation
}
