% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{group_similar_solutions}
\alias{group_similar_solutions}
\title{Group similar solutions for batch processing}
\usage{
group_similar_solutions(population, target_batch_size)
}
\arguments{
\item{population}{List of solutions}

\item{target_batch_size}{Target size for each batch}
}
\value{
List of solution groups
}
\description{
Group similar solutions for batch processing
}
