% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{i_optimality}
\alias{i_optimality}
\title{I-optimality criterion (minimizes average prediction variance)}
\usage{
i_optimality(train, test = NULL, P, lambda = 1e-06)
}
\arguments{
\item{train}{Vector of training set individual names}

\item{test}{Vector of test set individual names (if NULL, uses training set)}

\item{P}{Prediction matrix with individuals as rows and variables as columns}

\item{lambda}{Ridge regularization parameter (default: 1e-6)}
}
\value{
I-optimality value
}
\description{
I-optimality criterion (minimizes average prediction variance)
}
\details{
I-optimality minimizes the average prediction variance over the design space.
}
\references{
Fedorov, V.V. (1972). Theory of Optimal Experiments.
}
