% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiobjective.R
\name{initialize_population}
\alias{initialize_population}
\title{Initialize random population}
\usage{
initialize_population(candidates, ntoselect, npop)
}
\arguments{
\item{candidates}{Candidate individuals}

\item{ntoselect}{Number to select per solution}

\item{npop}{Population size}
}
\value{
List of initial solutions
}
\description{
Initialize random population
}
