% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiobjective.R
\name{multiobjective_selection}
\alias{multiobjective_selection}
\title{Multi-objective selection based on dominance ranks}
\usage{
multiobjective_selection(fitness, ranks, n_select)
}
\arguments{
\item{fitness}{Fitness matrix}

\item{ranks}{Dominance ranks}

\item{n_select}{Number of solutions to select}
}
\value{
Indices of selected solutions
}
\description{
Multi-objective selection based on dominance ranks
}
