% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga-operators.R
\name{mutate_solution}
\alias{mutate_solution}
\title{Mutate a solution}
\usage{
mutate_solution(solution, candidates, intensity, ntoselect)
}
\arguments{
\item{solution}{Current solution to mutate}

\item{candidates}{Pool of candidates}

\item{intensity}{Mutation intensity}

\item{ntoselect}{Target solution size}
}
\value{
Mutated solution
}
\description{
Mutate a solution
}
