% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiobjective.R
\name{non_dominated_sorting}
\alias{non_dominated_sorting}
\title{Non-dominated sorting for multi-objective optimization}
\usage{
non_dominated_sorting(fitness, criteria_types)
}
\arguments{
\item{fitness}{Fitness matrix}

\item{criteria_types}{Vector indicating minimize (TRUE) or maximize (FALSE)}
}
\value{
Vector of dominance ranks
}
\description{
Non-dominated sorting for multi-objective optimization
}
