% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{parallel_apply}
\alias{parallel_apply}
\title{Parallel matrix operations with proper error handling}
\usage{
parallel_apply(X, FUN, mc.cores = parallel::detectCores() - 1, ...)
}
\arguments{
\item{X}{List of matrices or data to process}

\item{FUN}{Function to apply}

\item{mc.cores}{Number of cores to use}

\item{...}{Additional arguments to FUN}
}
\value{
List of results
}
\description{
Parallel matrix operations with proper error handling
}
