\name{PEVMEANMM}
\alias{PEVMEANMM}

\title{PEVMEANMM}

\description{Kinv is a symmetric matrix (inverse of the relationship matrix K  between individuals). \eqn{Z_{Test}} and \eqn{Z_{Train}} are the design matrices for the individuals in the test and training sets correspondingly. M is calculated as \eqn{1_{ntrain}(1_{ntrain}'1_{ntrain})^{-1}1_{ntrain}'.}
Calculate  \deqn{mean(diag(CZ_{Test}(Z_{Train}'MZ_{Train}+\lambda*Kinv)^{-1}Z_{Test}'C'))).} C is taken as identity matrix if not specified. 
}
\usage{
PEVMEANMM(Train, Test, Kinv,K=NULL, lambda = 1e-05, C=NULL, Vg=NULL, Ve=NULL)
}
\arguments{
  \item{Train}{
vector of identifiers for individuals in the training set
}
  \item{Test}{
vector of identifiers for individuals in the test set
}
  \item{Kinv}{
\eqn{n \times n} inverse of the relationship matrix for the individuals with row and column names.
}
  \item{K}{
\eqn{n \times n} the relationship matrix for the individuals with row and column names.
}
  \item{lambda}{
scalar shrinkage parameter (\eqn{\lambda>0}).
}
 \item{C}{
Contrast Matrix.
}
 \item{Vg}{
covariance matrix between traits generated by the relationship K (multi-trait version).
}
 \item{Ve}{
residual covariance matrix for the traits (multi-trait version). 
}
}

\value{
value of the criterion.
}

\author{
Deniz Akdemir
}
