% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiobjective.R
\name{plot_pareto_progress}
\alias{plot_pareto_progress}
\title{Plot Pareto front progress}
\usage{
plot_pareto_progress(fitness, criteria, directions, generation, axes_labels)
}
\arguments{
\item{fitness}{Fitness matrix}

\item{criteria}{Criteria names}

\item{directions}{Plot directions}

\item{generation}{Current generation}

\item{axes_labels}{Axis labels}
}
\description{
Plot Pareto front progress
}
