% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{population_distances}
\alias{population_distances}
\title{Calculate population diversity metrics}
\usage{
population_distances(solutions, method = c("hamming", "jaccard", "sorensen"))
}
\arguments{
\item{solutions}{List of solutions in the population}

\item{method}{Distance calculation method: "hamming", "jaccard", "sorensen" (default: "jaccard")}
}
\value{
Distance matrix or diversity summary statistics
}
\description{
Calculate population diversity metrics
}
