% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiobjective.R
\name{select_diverse_subset}
\alias{select_diverse_subset}
\title{Select diverse subset using crowding distance}
\usage{
select_diverse_subset(fitness, max_size)
}
\arguments{
\item{fitness}{Fitness matrix}

\item{max_size}{Maximum number of solutions to select}
}
\value{
Indices of selected solutions
}
\description{
Select diverse subset using crowding distance
}
