% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiobjective.R
\name{subset_ga_multiobjective_single}
\alias{subset_ga_multiobjective_single}
\title{Multi-objective GA without test set (single population)}
\usage{
subset_ga_multiobjective_single(
  Pcs = NULL,
  Dist = NULL,
  Kernel = NULL,
  candidates,
  ntoselect,
  criteria,
  criteria_types,
  plot_directions,
  ...
)
}
\arguments{
\item{Pcs}{Principal components matrix}

\item{Dist}{Distance matrix}

\item{Kernel}{Kernel matrix}

\item{candidates}{Candidate individuals}

\item{ntoselect}{Number to select}

\item{criteria}{Optimization criteria}

\item{criteria_types}{Minimize/maximize indicators}

\item{plot_directions}{Plot directions}

\item{...}{Additional parameters}
}
\value{
Multi-objective optimization results
}
\description{
Multi-objective GA without test set (single population)
}
