% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga-operators.R
\name{tournament_selection}
\alias{tournament_selection}
\title{Tournament selection}
\usage{
tournament_selection(population, fitness, tournament_size = 3, n_select = 1)
}
\arguments{
\item{population}{List of solutions}

\item{fitness}{Vector of fitness values}

\item{tournament_size}{Size of tournament (default: 3)}

\item{n_select}{Number of individuals to select (default: 1)}
}
\value{
Indices of selected individuals
}
\description{
Tournament selection
}
