% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{transform_fitness}
\alias{transform_fitness}
\title{Fitness transformation and normalization utilities}
\usage{
transform_fitness(fitness, method = "linear", scaling_factor = 2)
}
\arguments{
\item{fitness}{Vector of fitness values}

\item{method}{Transformation method: "linear", "rank", "tournament" (default: "linear")}

\item{scaling_factor}{Scaling factor for transformations (default: 2)}
}
\value{
Transformed fitness values
}
\description{
Fitness transformation and normalization utilities
}
