% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiobjective.R
\name{update_pareto_archive}
\alias{update_pareto_archive}
\title{Update Pareto archive with new solutions}
\usage{
update_pareto_archive(
  population,
  fitness,
  archive,
  archive_fitness,
  max_size,
  criteria_types
)
}
\arguments{
\item{population}{Current population}

\item{fitness}{Fitness matrix}

\item{archive}{Current archive}

\item{archive_fitness}{Archive fitness matrix}

\item{max_size}{Maximum archive size}

\item{criteria_types}{Optimization types (minimize/maximize)}
}
\value{
Updated archive and fitness
}
\description{
Update Pareto archive with new solutions
}
