% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{update_progress}
\alias{update_progress}
\title{Update progress bar}
\usage{
update_progress(pb, increment = 1)
}
\arguments{
\item{pb}{Progress bar object}

\item{increment}{How much to increment (default: 1)}
}
\description{
Update progress bar
}
