% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_kinship_matrix}
\alias{validate_kinship_matrix}
\title{Validate kinship matrix}
\usage{
validate_kinship_matrix(K, individual_names)
}
\arguments{
\item{K}{Kinship matrix}

\item{individual_names}{Names of individuals to check against}
}
\value{
NULL if valid, throws error if invalid
}
\description{
Validate kinship matrix
}
