% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R, R/criteria_old.R
\name{h2_to_variances}
\alias{h2_to_variances}
\title{STPGA Optimization Criteria Functions}
\usage{
h2_to_variances(h2, total_var = 1)

h2_to_variances(h2, total_var = 1)
}
\arguments{
\item{h2}{Heritability (proportion of variance due to genetics)}

\item{total_var}{Total variance (default: 1)}
}
\value{
List with Vg (genetic variance) and Ve (error variance)

List with Vg (genetic variance) and Ve (error variance)
}
\description{
Literature-corrected optimization criteria for subset selection in experimental design

Optimization criteria corrected according to published experimental design literature
}
\references{
Fedorov, V.V. (1972). Theory of Optimal Experiments. Academic Press.
Henderson, C.R. (1984). Applications of Linear Models in Animal Breeding.
Kiefer, J. (1959). Optimum experimental designs. Journal of the Royal Statistical Society B.
Searle, S.R., Casella, G., McCulloch, C.E. (1992). Variance Components. Wiley.
Atkinson, A.C., Donev, A.N., Tobias, R.D. (2007). Optimum Experimental Designs. Oxford.
Convert between variance components and heritability

Fedorov, V.V. (1972). Theory of Optimal Experiments. Academic Press.
Henderson, C.R. (1984). Applications of Linear Models in Animal Breeding.
Kiefer, J. (1959). Optimum experimental designs. Journal of the Royal Statistical Society B.
Searle, S.R., Casella, G., McCulloch, C.E. (1992). Variance Components. Wiley.
Atkinson, A.C., Donev, A.N., Tobias, R.D. (2007). Optimum Experimental Designs. Oxford.
Convert between variance components and heritability
}
\author{
Deniz Akdemir
}
